requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pyobjc-framework-Cocoa>=10.3
pyobjc-framework-Quartz>=10.3
pyobjc-framework-ApplicationServices>=10.3
httpx>=0.27.0
cryptography>=42.0
qrcode>=7.4
websockets>=12.0
PyJWT>=2.10.0
